Kernel: Call TARGET_OUT_INTERMEDIATES from top of source code
On build, kernel.mk would try to access the output KERNEL_OBJ path from its current/local directory, rather than from the top of the android source code.

Adding $PWD to the path ensures that it gets read from the current directory in which a make has been executed, which may only be done at the top of the android
source.

Signed-off-by: CTXz <ctx.xda@gmail.com>